@@ -11,57 +11,58 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.10",
+    "@angular/animations": "^8.0.0",
     "@angular/cdk": "^7.3.7",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/core": "^6.1.0",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
-    "@angular/material": "^7.2.1",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/router": "^6.1.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "^8.0.0",
+    "@angular/material": "^7.3.7",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
     "@asymmetrik/ngx-leaflet": "^4.0.0",
-    "@ng-select/ng-select": "^2.14.0",
-    "@ngx-gallery/core": "^4.0.2-beta.0",
-    "@ngx-gallery/lightbox": "^4.0.2-beta.0",
+    "@ng-select/ng-select": "^2.20.0",
+    "@ngx-gallery/core": "^4.0.3",
+    "@ngx-gallery/lightbox": "^4.0.3",
     "@ngx-translate/core": "^10.0.2",
-    "core-js": "^2.5.4",
-    "date-fns": "^1.29.0",
-    "leaflet": "^1.3.4",
+    "core-js": "^2.6.9",
+    "date-fns": "^1.30.1",
+    "leaflet": "^1.5.1",
     "material-design-icons": "^3.0.1",
-    "ngx-avatar": "^3.2.0",
+    "ngx-avatar": "^3.7.0",
     "ngx-filter-pipe": "^2.1.2",
-    "ngx-input-file": "^7.0.1",
+    "ngx-input-file": "^7.0.4",
     "ngx-lightbox": "^1.2.0",
     "ngx-modal": "0.0.29",
     "ngx-owl-carousel": "^2.0.7",
-    "ngx-pipes": "^2.3.6",
+    "ngx-pipes": "^2.5.1",
+    "ngx-smart-modal": "^7.1.1",
     "pretty-checkbox": "^3.0.3",
-    "rxjs": "^6.0.0",
-    "zone.js": "~0.8.26"
+    "rxjs": "^6.5.2",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.7.0",
-    "@angular/cli": "~6.1.2",
-    "@angular/compiler-cli": "^6.1.0",
-    "@angular/language-service": "^6.1.0",
-    "@types/jasmine": "~2.8.6",
-    "@types/jasminewd2": "~2.0.3",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular/cli": "~8.0.2",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "^8.0.0",
+    "@types/jasmine": "^2.8.16",
+    "@types/jasminewd2": "^2.0.6",
     "@types/jquery": "^3.3.29",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~1.7.1",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
+    "karma-coverage-istanbul-reporter": "^2.0.5",
     "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.3.0",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "typescript": "~3.4.5"
   }
 }
