@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
-import { Lightbox } from 'ngx-lightbox';
+import {Gallery, ImageItem, VideoItem} from '@ngx-gallery/core';
+
 
 @Component({
   selector: 'app-ad-view',
@@ -14,16 +15,8 @@ export class AdViewComponent implements OnInit {
   toggle2 = true;
   public _albums= [];
 
-  constructor(private route: ActivatedRoute, private tr: TranslateService, private _lightbox: Lightbox) {}
-  open(index: number): void {
-    // open lightbox
-    this._lightbox.open(this._albums, index);
-  }
+  constructor(private gallery: Gallery ,private route: ActivatedRoute, private tr: TranslateService) {}
 
-  close(): void {
-    // close lightbox programmatically
-    this._lightbox.close();
-  }
   ngOnInit() {
     this.route.data.subscribe(({adData}) =>{
       // console.warn (adData);
@@ -33,20 +26,24 @@ export class AdViewComponent implements OnInit {
       this.data['description'] = adData['description'];
       var t = new Date(adData['creationDate']);
       this.data['creationDate'] = t.toLocaleDateString() + '-' + t.toLocaleTimeString();
-      // this.data['images']= adData['media']; 
+      // this.data['images']= adData['media'];
       for (let i = 0; i < adData['media'].length; i++) {
         const src = adData['media'][i]['url'];
         const caption = 'Image' + adData['media'][i]['id']  ;
         const thumb = adData['media'][i]['thumbnail'];
         const album = {
           src: src,
-          caption: caption,
+          title: caption,
           thumb: thumb
         };
-
-        this._albums.push(album);
+        if(adData['media'][i]['type']=='image')
+          this._albums.push(new ImageItem(album));
+        else
+          this._albums.push(new VideoItem(album));
+        // this._albums.push(album);
       }
-
+      const galleryRef = this.gallery.ref();
+      galleryRef.load(this._albums)
       if (this.tr.currentLang=='ar'){
         this.data['area']=adData['city']['nameAr'];
         this.data['subArea']=adData['location']['nameAr'];
